+++ lib/pages/track/widgets/track_details.dart
@@ -1,76 +1,25 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:duration/duration.dart';
 import 'package:flutter/material.dart';
-import 'package:palette_generator/palette_generator.dart';
-import 'package:scoped_model/scoped_model.dart';
 import 'package:spotify/spotify_io.dart' as spotify;
-import 'package:url_launcher/url_launcher.dart';
 
-import 'package:tasty_tracks/models/track_bet_model.dart';
-import 'package:tasty_tracks/models/track_watch_model.dart';
-import 'package:tasty_tracks/pages/album/album.dart';
 import 'package:tasty_tracks/pages/album/widgets/album_image.dart';
-import 'package:tasty_tracks/pages/artist/artist.dart';
-import 'package:tasty_tracks/pages/track/widgets/track_preview_player.dart';
-import 'package:tasty_tracks/pages/track_bet/create.dart';
 import 'package:tasty_tracks/utils/format_date.dart';
-
-enum MenuActions { viewAlbum, viewArtist, openSpotify }
 
 class TrackDetails extends StatelessWidget {
   TrackDetails({
     Key key,
     this.track,
     this.album,
-    this.palette,
     this.heroSuffix,
   }) : super(key: key);
 
   final spotify.Track track;
   final spotify.Album album;
-  final PaletteGenerator palette;
   final String heroSuffix;
-
-  final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
 
   @override
   Widget build(BuildContext context) {
     ThemeData theme = Theme.of(context);
-
-    Widget appBar = AppBar(
-      backgroundColor: theme.canvasColor,
-      actions: <Widget>[
-        PopupMenuButton(
-          onSelected: (MenuActions result) {
-            switch (result) {
-              case MenuActions.viewAlbum:
-                _viewAlbum(context);
-                break;
-              case MenuActions.viewArtist:
-                _viewArtist(context);
-                break;
-              case MenuActions.openSpotify:
-                _openSpotify(context);
-                break;
-            }
-          },
-          itemBuilder: (BuildContext context) => <PopupMenuEntry<MenuActions>>[
-                PopupMenuItem<MenuActions>(
-                  value: MenuActions.viewAlbum,
-                  child: Text('View album'),
-                ),
-                PopupMenuItem(
-                  value: MenuActions.viewArtist,
-                  child: Text('View artist'),
-                ),
-                PopupMenuItem(
-                  value: MenuActions.openSpotify,
-                  child: Text('Open with Spotify'),
-                ),
-              ],
-        ),
-      ],
-    );
 
     String artistNames = track.artists.map((artist) => artist.name).join(', ');
 
@@ -84,14 +33,14 @@ class TrackDetails extends StatelessWidget {
       ),
     );
 
-    Widget header = Container(
+    return Container(
       padding: EdgeInsets.symmetric(horizontal: 32.0, vertical: 16.0),
       decoration: BoxDecoration(
         gradient: LinearGradient(
           begin: Alignment.topCenter,
           end: Alignment.bottomCenter,
           colors: [
-            theme.accentColor.withOpacity(0.4),
+            theme.accentColor.withOpacity(0.25),
             theme.accentColor.withOpacity(0.0)
           ],
         ),
@@ -133,179 +82,5 @@ class TrackDetails extends StatelessWidget {
         ],
       ),
     );
-
-    Widget previewPlayer = TrackPreviewPlayer(previewUrl: track.previewUrl);
-
-    Widget popularity = Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        Text(
-          '${track.popularity}',
-          style: theme.textTheme.display2.copyWith(color: theme.accentColor),
-        ),
-        const SizedBox(width: 4.0),
-        Text(
-          '/100 popularity',
-          style: theme.textTheme.subhead,
-        ),
-      ],
-    );
-
-    Widget buttons = Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: <Widget>[
-        ScopedModelDescendant<TrackBetModel>(
-          rebuildOnChange: false,
-          builder: (context, child, trackBetModel) {
-            return StreamBuilder(
-              stream: trackBetModel.snapshots(trackId: track.id),
-              builder: (BuildContext context,
-                  AsyncSnapshot<QuerySnapshot> snapshot) {
-                if (snapshot.hasData && snapshot.data.documents.isNotEmpty) {
-                  return RaisedButton(
-                    child: Text('Remove bet'),
-                    color: theme.accentColor,
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(64.0),
-                    ),
-                    onPressed: () => trackBetModel.remove(track.id),
-                  );
-                } else {
-                  return RaisedButton(
-                    child: Text('Place Bet'),
-                    color: theme.accentColor,
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(64.0),
-                    ),
-                    // TODO Show bet dialog
-                    onPressed: () => _placeBet(context),
-                  );
-                }
-              },
-            );
-          },
-        ),
-        ScopedModelDescendant<TrackWatchModel>(
-          rebuildOnChange: false,
-          builder: (context, child, trackWatchModel) {
-            return StreamBuilder(
-              stream: trackWatchModel.snapshots(trackId: track.id),
-              builder: (BuildContext context,
-                  AsyncSnapshot<QuerySnapshot> snapshot) {
-                if (snapshot.hasData && snapshot.data.documents.isNotEmpty) {
-                  return OutlineButton(
-                    child: Text('Remove watch'),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(64.0),
-                    ),
-                    onPressed: () => trackWatchModel.remove(track.id),
-                  );
-                } else {
-                  return OutlineButton(
-                    child: Text('Watch'),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(64.0),
-                    ),
-                    onPressed: () => trackWatchModel.add(track.id),
-                  );
-                }
-              },
-            );
-          },
-        ),
-      ],
-    );
-
-    Widget copyrights = Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: album.copyrights.map((c) {
-        return Text(
-          c.text,
-          style: theme.textTheme.caption,
-        );
-      }).toList(),
-    );
-
-    Widget credits = Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Text('Available in ${album.availableMarkets.length} markets.'),
-        const SizedBox(height: 8.0),
-        Text('${album.label}'),
-        const SizedBox(height: 4.0),
-        copyrights,
-      ],
-    );
-
-    Widget body = Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 32.0),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: <Widget>[
-          popularity,
-          const SizedBox(height: 8.0),
-          buttons,
-          const SizedBox(height: 8.0),
-          previewPlayer,
-          const SizedBox(height: 16.0),
-          credits,
-        ],
-      ),
-    );
-
-    return Scaffold(
-      key: _scaffoldKey,
-      appBar: appBar,
-      body: SafeArea(
-        child: ListView(
-          padding: EdgeInsets.only(
-            bottom: 64.0,
-          ),
-          children: <Widget>[
-            header,
-            body,
-          ],
-        ),
-      ),
-    );
-  }
-
-  _placeBet(BuildContext context) {
-    Navigator.of(context).pushNamed(
-      TrackBetCreate.routeName,
-      arguments: {
-        'track_id': track.id,
-      },
-    );
-  }
-
-  _viewAlbum(BuildContext context) {
-    Navigator.of(context).pushNamed(
-      AlbumPage.routeName,
-      arguments: {
-        'album_id': track.album.id,
-      },
-    );
-  }
-
-  _viewArtist(BuildContext context) {
-    Navigator.of(context).pushNamed(
-      ArtistPage.routeName,
-      arguments: {
-        'artist_id': track.artists.first.id,
-      },
-    );
-  }
-
-  _openSpotify(BuildContext context) async {
-    String url = track.uri;
-
-    if (await canLaunch(url)) {
-      await launch(url);
-    } else {
-      String errorMessage = "Couldn't open with Spotify";
-      _scaffoldKey.currentState
-          .showSnackBar(SnackBar(content: Text(errorMessage)));
-    }
   }
 }
